requirements.txt
requests
numpy
pypdf
python-dotenv
chromadb
rank-bm25
scikit-learn
tqdm
pytest
faiss-cpu
qdrant-client